+++ src/app/modules/navBar.js
@@ -1,9 +1,8 @@
 "use client";
 
 import React, { useRef, useState, useEffect } from "react";
+import styled from "styled-components";
 import SocialIcons from "./socialIcons";
-import Menu from "./menu";
-import Burger from "./burger";
 
 const useOnClickOutside = (ref, handler) => {
   useEffect(() => {
@@ -23,14 +22,13 @@ const useOnClickOutside = (ref, handler) => {
   }, [ref, handler]);
 };
 
-const Navbar = () => {
+export default function Navbar () {
   const scrollToSection = (id) => {
     document.getElementById(id)?.scrollIntoView({ behavior: "smooth" });
   };
 
   const [open, setOpen] = useState(false);
   const node = useRef();
-
   useOnClickOutside(node, () => setOpen(false));
 
   return (
@@ -86,4 +84,118 @@ const Navbar = () => {
   );
 };
 
-export default Navbar;
+const StyledBurger = styled.button`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  width: 2rem;
+  height: 2rem;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  padding: 0;
+  z-index: 10;
+
+  &:focus {
+    outline: none;
+  }
+
+  div {
+    width: 2rem;
+    height: 0.25rem;
+    background: ${({ open }) => (open ? "#ffa62b" : "#ede7e4")};
+    border-radius: 10px;
+    transition: all 0.3s linear;
+    position: relative;
+    transform-origin: 1px;
+  }
+
+  :first-child {
+    transform: ${({ open }) => (open ? "rotate(45deg)" : "rotate(0)")};
+  }
+
+  :nth-child(2) {
+    opacity: ${({ open }) => (open ? "0" : "1")};
+    transform: ${({ open }) => (open ? "translateX(20px)" : "translateX(0)")};
+  }
+
+  :nth-child(3) {
+    transform: ${({ open }) => (open ? "rotate(-45deg)" : "rotate(0)")};
+  }
+`;
+
+const Burger = ({ open, setOpen }) => {
+  return (
+    <StyledBurger open={open} onClick={() => setOpen(!open)} className="hover:text-blue-dark">
+      <div />
+      <div />
+      <div />
+    </StyledBurger>
+  );
+};
+
+const StyledMenu = styled.nav`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  background: #ede7e4;
+  transform: ${({ open }) => (open ? "translateX(0)" : "translateX(-100%)")};
+  width: 100vw;
+  height: 100vh;
+  text-align: left;
+  padding: 2rem;
+  position: absolute;
+  top: 0;
+  left: 0;
+  transition: transform 0.3s ease-in-out;
+
+  button {
+    font-size: 1.4rem;
+    text-align: center;
+    text-transform: uppercase;
+    padding: 2rem 0;
+    font-weight: bold;
+    letter-spacing: 0.5rem;
+    color: #0d0c1d;
+    text-decoration: none;
+    transition: color 0.3s linear;
+    color: #16697a;
+  }
+`;
+
+const Menu = ({ open, setOpen }) => {
+  const scrollToSection = (id) => {
+    const section = document.getElementById(id);
+    const navbarHeight = document.getElementById('navbar')?.offsetHeight || 0;
+    setOpen(false);
+
+    if (section) {
+      window.scrollTo({
+        top: section.offsetTop - navbarHeight,
+        behavior: "smooth",
+      });
+    }
+  };
+  return (
+    <StyledMenu open={open}>
+      <button onClick={() => scrollToSection("servizi")}>
+        <span role="img" aria-label="about us">
+          💸
+        </span>
+        Servizi
+      </button>
+      <button onClick={() => scrollToSection("aboutUs")}>
+        <span role="img" aria-label="price">
+          💁🏻‍♂️
+        </span>
+        Chi Siamo
+      </button>
+      <button onClick={() => scrollToSection("contattaci")}>
+        <span role="img" aria-label="contact">
+          📩
+        </span>
+        Contattaci
+      </button>
+    </StyledMenu>
+  );
+};
